backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
requests>=2.31
python-dotenv>=1.0
chromadb>=0.5
sentence-transformers>=3.0
